Fix mismatchin in byte size of step number

+++ plot/login.py
@@ -74,7 +74,7 @@ if __name__ == "__main__":
         setup[name]['io'] = adios.DeclareIO(name)
         #setup[name]['io'].SetEngine('BP4')
         setup[name]['opened'] = False
-        setup[name]['LastStep'] = np.array([-1], dtype=np.int64)
+        setup[name]['LastStep'] = np.array([-1], dtype=np.int32)
         setup[name]['done'] = False
